Fix VLAN interface re-initialization after driver reload
When we set up a VLAN interface using NMState or nmcli on top of a NIC
managed by the ice driver, the process of removing the in-tree driver
and loading the out-of-tree driver will introduce a delay before VLAN
interfaces are re-enabled. This is an issue when that VLAN interface is
the primary NIC as seen by the OpenShift node.

This commit adds code to the ptp-config.sh script, where we can
configure the VLAN interfaces we have in our configuration. This allows
us to re-enable the interfaces after the out-of-tree driver has been
loaded, and gives the configure-ovs.sh script a chance to find the right
external IP.